Maybe now ?

+++ src/main/resources/init_db.sql
@@ -1,5 +1,5 @@
 CREATE TABLE books (
-    id INT NOT NULL AUTO_INCREMENT,
+    id LONG NOT NULL AUTO_INCREMENT,
     title VARCHAR(255),
     price FLOAT,
     PRIMARY KEY (id)
